finally broke hierarchy_vis into a js module
