+++ snowflake-inspector/src/css/style.css
@@ -1,8 +1,7 @@
-#mynetwork {
+#hierarchy-vis {
   width: 100%;
   height: 100vh;
   overflow: hidden;
-  /* border: 1px solid lightgray; */
 }
 .queryStyle {
   height: 300px;     
@@ -11,8 +10,4 @@
   border: 1px; 
   background-color: rgb(211,211,211, 0.3); 
   border-radius: 5px;
-}
-
-html {
-  background-color: red;
 }
