+++ snowflake-inspector/src/html/index.html
@@ -37,93 +37,8 @@
 
   <!-- Begin page content -->
   <main role="main" class="flex-shrink-0">
-    <div class="container">
-      <!-- Modal -->
-      <div class="modal fade" id="createModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel"
-        aria-hidden="true">
-        <div class="modal-dialog modal-lg" role="document">
-          <div class="modal-content">
-            <div class="modal-header">
-              <h5 class="modal-title" id="exampleModalLabel">Run this query in your Snowflake account and paste the
-                result in the box below</h5>
-              <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                <span aria-hidden="true">&times;</span>
-              </button>
-            </div>
-            <div class="modal-body">
-              <div class="queryStyle">
-                <pre>
-                <code class="to-copy"> 
-                  USE ROLE ACCOUNTADMIN;
-                  WITH 
-                  ROLE_GRANTS AS (
-                    SELECT
-                      PRIVILEGE,
-                      GRANTED_ON AS SNOWFLAKE_OBJECT_TYPE,
-                      TABLE_CATALOG AS DATABASE,
-                      TABLE_SCHEMA AS SCHEMA,
-                      NAME AS OBJECT_NAME,
-                      GRANTED_TO AS GRANTED_TO_TYPE,
-                      GRANTEE_NAME AS GRANTED_TO_NAME
-                    FROM
-                      SNOWFLAKE.ACCOUNT_USAGE.GRANTS_TO_ROLES
-                    WHERE
-                      DELETED_ON IS NULL
-                  ),
-                  USER_GRANTS AS (
-                    SELECT
-                    'USAGE' AS PRIVILEGE,
-                    'ROLE' AS SNOWFLAKE_OBJECT_TYPE,
-                    NULL AS DATABASE,
-                    NULL AS SCHEMA,
-                    ROLE AS OBJECT_NAME,
-                    GRANTED_TO AS GRANTED_TO_TYPE,
-                    GRANTEE_NAME AS GRANTED_TO_NAME
-                    FROM
-                      SNOWFLAKE.ACCOUNT_USAGE.GRANTS_TO_USERS
-                    WHERE
-                      DELETED_ON IS NULL
-                  ),
-                  GRANTS AS (
-                    SELECT * FROM ROLE_GRANTS 
-                    UNION ALL 
-                    SELECT * FROM USER_GRANTS
-                  ),
-                  GRANTS_JSON AS (
-                    SELECT
-                      OBJECT_CONSTRUCT(*) AS GRANT_OBJ
-                    FROM
-                      GRANTS
-                  ),
-                  GRANTS_JSON_ARR AS (
-                    SELECT
-                      array_agg(*) AS GRANTS
-                    FROM
-                      GRANTS_JSON
-                  )
-                  SELECT * FROM GRANTS_JSON_ARR;
-                </code>
-            </pre>
-              </div>
-              <div>
-                <button type="button" class="btn btn-secondary btn-round write-btn">Copy to clipboard</button>
-              </div>
-              <div class="mt-4">
-                <form>
-                  <div class="form-group">
-                    <label>Paste the query output here</label>
-                    <textarea class="form-control" name="result" id="result" rows="5" required="true"></textarea>
-                  </div>
-                  <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary btn-round" data-dismiss="modal">Close</button>
-                    <button type="submit" class="btn btn-primary btn-round">Visualize</button>
-                  </div>
-                </form>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
+    <div class="container-fluid">
+      <div id="hierarchy-vis"></div>
     </div>
   </main>
